+++ src/SWP/Component/Bridge/Model/ItemInterface.php
@@ -13,8 +13,15 @@
  */
 namespace SWP\Component\Bridge\Model;
 
+/**
+ * Interface ItemInterface
+ */
 interface ItemInterface extends ContentInterface
 {
+    const TYPE_TEXT = 'text';
+    const TYPE_FILE = 'file';
+    const TYPE_PICTURE = 'picture';
+
     /**
      * @return string
      */